Update scripts and plan
